Fixed crasher in drawSphere call to collides.
We ensure that every pixel we draw is within the bounds of the frame,
but were sampling the depth map at the sphere's center. This has been
switched to a per-pixel depth test that benefits from the existing
bounds-checking. An alternative to possibly improve rendering
usefulness would be to render a sphere red if *any* pixel collides,
and green otherwise. This could give a good signal to the sure.