Added a build flag to disable mesh loading.
The required data file (newpr2.txt) is not in the repository, causing
bizarre runtime errors due to a lack of sufficient error handling in
model loading. DAE loading also depends on the TinyXML library, which
is not currently available via homebrew, so a bit of a hassle when not
building with rosmake. The new DAE constant enables the loading of DAE
meshes, and so is currently undefined for mac builds.

A memory leak and logic oddness were fixed in Joint.h.