Vector interface to CollisionChecker.
The vector interface is much faster than the reference implementation
that uses maps. To make usage more straightforward, CollisionChecker
now has helper functions to convert a map of joint angles to a vector
of joint angle assignments, and to convert a vector of collision
information to a map keyed by joint name.