Bring camera transformations into harmony.
Change coordinate frame transformations to be consistent with the
point cloud viewer used to generate depth maps. Among several changes,
a significant alteration is that the Y-axis is inverted when going
from camera space to screen space. This is the OpenGL convention,
which we now adhere to.

There are also now two focal lengths: one for each screen space axis.

Keyboard controls of the robot have been rationalized so it can be
guided around the environment.