Program for pre-rendering spheres.
The pre-rendered spheres may be used by the depth map dilation
procedure to reduce computation. The idea is that we cache spheres of
every possible integer radius so that dilating a point involves
looking up the appropriately-sized cached rendering, then scaling
those depths by the actual sphere radius and comparing with the
destination depth map.